requirements.txt
python-telegram-bot
supabase
flask
python-dotenv